Refactor: layout - add header in main layout(extract ScrollHeader) - modify folder structre main layout - modify event detail view layout
