+++ src/app/(main)/ScrollHeader.tsx
@@ -16,7 +16,7 @@ const ScrollHeader: React.FC<HeaderProps> = ({ className = '', children }) => {
   return (
     <header
       className={`
-        fixed top-0 left-0 right-0 z-40
+        sticky top-0 left-0 right-0 z-40
         transition-transform duration-300 ease-in-out
         ${isVisible ? 'translate-y-0' : '-translate-y-full'}
         ${className}
