+++ src/components/layout/MainLayout.tsx
@@ -1,4 +1,5 @@
 import { Geist, Geist_Mono } from 'next/font/google';
+import Header from '../common/Header';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -12,13 +13,15 @@ const geistMono = Geist_Mono({
 
 export default function MainLayout({
   children,
+  isStory = false,
 }: Readonly<{
   children: React.ReactNode;
-}>) {
+}> & { isStory?: boolean }) {
   return (
     <div
       className={`${geistSans.variable} ${geistMono.variable} antialiased w-full min-h-screen bg-gradient-to-br from-slate-50 via-white to-indigo-50`}
     >
+      {!isStory && <Header />}
       {children}
     </div>
   );
